Refactor city and country modules with model, service, and exception handling updates
- Added CityController, CityService, CityModel, CountryModel, and ResourceNotFoundException
- Updated CountryController, CountryService, and CountryRepository
- Refactored City and BaseEntity entities
- Enhanced GlobalExceptionHandler and utility constants
- Removed CountryRequestDto and cleaned up related code
- Updated pom.xml configurations

+++ src/main/java/com/example/Flight/Booking/System/entity/City.java
@@ -20,5 +20,5 @@ public class City extends BaseEntity {
 	private String cityName;
 
 	@Column(name = "country_id", nullable = false)
-	private Integer countryId;
+	private Long countryId;
 }
